refactor(backend): move workflow create update commands
