+++ backend/src/main/java/com/onedata/portal/service/WorkflowCommandService.java
@@ -1,19 +1,33 @@
 package com.onedata.portal.service;
 
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+import com.onedata.portal.dto.workflow.WorkflowDefinitionRequest;
+import com.onedata.portal.dto.workflow.WorkflowTaskBinding;
+import com.onedata.portal.dto.workflow.WorkflowTopologyResult;
 import com.onedata.portal.entity.DataLineage;
 import com.onedata.portal.entity.DataWorkflow;
+import com.onedata.portal.entity.DolphinConfig;
 import com.onedata.portal.entity.TableTaskRelation;
 import com.onedata.portal.entity.WorkflowTaskRelation;
+import com.onedata.portal.entity.WorkflowVersion;
 import com.onedata.portal.mapper.DataLineageMapper;
 import com.onedata.portal.mapper.DataTaskMapper;
 import com.onedata.portal.mapper.DataWorkflowMapper;
 import com.onedata.portal.mapper.TableTaskRelationMapper;
 import com.onedata.portal.mapper.WorkflowTaskRelationMapper;
+import com.onedata.portal.mapper.WorkflowVersionMapper;
+import com.onedata.portal.util.JsonCanonicalizer;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
+import org.springframework.util.StringUtils;
 
+import java.time.LocalDateTime;
+import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
@@ -26,12 +40,106 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class WorkflowCommandService {
 
+    private static final int SNAPSHOT_SCHEMA_VERSION_DEFINITION = 3;
+
     private final DataWorkflowMapper dataWorkflowMapper;
     private final WorkflowTaskRelationMapper workflowTaskRelationMapper;
     private final DolphinSchedulerService dolphinSchedulerService;
     private final DataTaskMapper dataTaskMapper;
     private final DataLineageMapper dataLineageMapper;
     private final TableTaskRelationMapper tableTaskRelationMapper;
+    private final WorkflowVersionService workflowVersionService;
+    private final WorkflowVersionMapper workflowVersionMapper;
+    private final ObjectMapper objectMapper;
+    private final WorkflowTopologyService workflowTopologyService;
+    private final DolphinConfigService dolphinConfigService;
+    private final WorkflowTaskRelationService workflowTaskRelationService;
+    private final WorkflowDefinitionAssembler workflowDefinitionAssembler;
+
+    public DataWorkflow createWorkflow(WorkflowDefinitionRequest request) {
+        DataWorkflow workflow = new DataWorkflow();
+        LocalDateTime now = LocalDateTime.now();
+        List<WorkflowTaskBinding> taskBindings = normalizeTaskBindings(request.getTasks());
+        request.setTasks(taskBindings);
+        List<Long> taskIdsInOrder = workflowTaskRelationService.collectTaskIds(taskBindings);
+        WorkflowTopologyResult topology = workflowTopologyService.buildTopology(taskIdsInOrder);
+        workflow.setWorkflowName(request.getWorkflowName());
+        workflow.setDescription(request.getDescription());
+        workflow.setDefinitionJson(workflowDefinitionAssembler.defaultJson(request.getDefinitionJson()));
+        workflow.setEntryTaskIds(workflowDefinitionAssembler.toJson(
+                workflowDefinitionAssembler.orderTaskIds(topology.getEntryTaskIds(), taskIdsInOrder)));
+        workflow.setExitTaskIds(workflowDefinitionAssembler.toJson(
+                workflowDefinitionAssembler.orderTaskIds(topology.getExitTaskIds(), taskIdsInOrder)));
+        workflow.setGlobalParams(request.getGlobalParams());
+        workflow.setTaskGroupName(request.getTaskGroupName());
+        workflow.setStatus("draft");
+        workflow.setPublishStatus("never");
+        workflow.setDolphinConfigId(resolveDolphinConfigId(request.getDolphinConfigId()));
+        workflow.setProjectCode(resolveProjectCode(request.getProjectCode()));
+        workflow.setCreatedBy(request.getOperator());
+        workflow.setUpdatedBy(request.getOperator());
+        workflow.setCreatedAt(now);
+        workflow.setUpdatedAt(now);
+        workflowDefinitionAssembler.normalizeWorkflowScheduleDefaults(workflow);
+        dataWorkflowMapper.insert(workflow);
+
+        workflowTaskRelationService.persistTaskRelations(workflow.getId(), taskBindings, null, topology);
+        workflowDefinitionAssembler.normalizeTaskMetadata(taskIdsInOrder, workflow.getTaskGroupName());
+
+        String resolvedDefinitionJson = workflowDefinitionAssembler.resolveDefinitionJson(workflow, request, taskBindings, topology);
+        workflow.setDefinitionJson(resolvedDefinitionJson);
+        dataWorkflowMapper.updateById(workflow);
+
+        WorkflowVersion version = snapshotWorkflow(workflow, request, resolvedDefinitionJson);
+        workflow.setCurrentVersionId(version.getId());
+        dataWorkflowMapper.updateById(workflow);
+
+        updateRelationVersion(workflow.getId(), version.getId());
+        return workflow;
+    }
+
+    public DataWorkflow updateWorkflow(Long workflowId, WorkflowDefinitionRequest request) {
+        DataWorkflow workflow = dataWorkflowMapper.selectById(workflowId);
+        if (workflow == null) {
+            throw new IllegalArgumentException("Workflow not found: " + workflowId);
+        }
+        List<WorkflowTaskBinding> taskBindings = normalizeTaskBindings(request.getTasks());
+        request.setTasks(taskBindings);
+        List<Long> taskIdsInOrder = workflowTaskRelationService.collectTaskIds(taskBindings);
+        WorkflowTopologyResult topology = workflowTopologyService.buildTopology(taskIdsInOrder);
+        workflow.setWorkflowName(request.getWorkflowName());
+        workflow.setDescription(request.getDescription());
+        workflow.setEntryTaskIds(workflowDefinitionAssembler.toJson(
+                workflowDefinitionAssembler.orderTaskIds(topology.getEntryTaskIds(), taskIdsInOrder)));
+        workflow.setExitTaskIds(workflowDefinitionAssembler.toJson(
+                workflowDefinitionAssembler.orderTaskIds(topology.getExitTaskIds(), taskIdsInOrder)));
+        workflow.setGlobalParams(request.getGlobalParams());
+        workflow.setTaskGroupName(request.getTaskGroupName());
+        workflow.setUpdatedBy(request.getOperator());
+        workflow.setUpdatedAt(LocalDateTime.now());
+        if (workflow.getDolphinConfigId() == null) {
+            workflow.setDolphinConfigId(resolveDolphinConfigId(request.getDolphinConfigId()));
+        }
+        if (workflow.getProjectCode() == null || workflow.getProjectCode() == 0) {
+            workflow.setProjectCode(resolveProjectCode(request.getProjectCode()));
+        }
+        workflowDefinitionAssembler.normalizeWorkflowScheduleDefaults(workflow);
+
+        workflowTaskRelationService.persistTaskRelations(workflowId, taskBindings, workflow.getCurrentVersionId(), topology);
+        workflowDefinitionAssembler.normalizeTaskMetadata(taskIdsInOrder, workflow.getTaskGroupName());
+
+        String resolvedDefinitionJson = workflowDefinitionAssembler.resolveDefinitionJson(workflow, request, taskBindings, topology);
+        workflow.setDefinitionJson(resolvedDefinitionJson);
+        dataWorkflowMapper.updateById(workflow);
+
+        if (shouldCreateNewVersion(workflow, resolvedDefinitionJson)) {
+            WorkflowVersion version = snapshotWorkflow(workflow, request, resolvedDefinitionJson);
+            workflow.setCurrentVersionId(version.getId());
+            dataWorkflowMapper.updateById(workflow);
+            updateRelationVersion(workflowId, version.getId());
+        }
+        return workflow;
+    }
 
     public void deleteWorkflow(Long workflowId, boolean cascadeDeleteTasks) {
         if (workflowId == null) {
@@ -102,5 +210,90 @@ public class WorkflowCommandService {
             log.error("删除工作流失败: {}", workflowId, e);
             throw new RuntimeException("删除工作流失败: " + e.getMessage(), e);
         }
+    }
+
+    private void updateRelationVersion(Long workflowId, Long versionId) {
+        WorkflowTaskRelation update = new WorkflowTaskRelation();
+        update.setVersionId(versionId);
+        workflowTaskRelationMapper.update(update,
+                Wrappers.<WorkflowTaskRelation>lambdaUpdate()
+                        .eq(WorkflowTaskRelation::getWorkflowId, workflowId));
+    }
+
+    private WorkflowVersion snapshotWorkflow(DataWorkflow workflow,
+            WorkflowDefinitionRequest request,
+            String snapshotJson) {
+        boolean isInitial = workflow.getCurrentVersionId() == null;
+        String changeSummary = isInitial ? "initial workflow definition" : "updated workflow definition";
+        return workflowVersionService.createVersion(
+                workflow.getId(),
+                snapshotJson,
+                StringUtils.hasText(request.getDescription()) ? request.getDescription() : changeSummary,
+                request.getTriggerSource(),
+                request.getOperator(),
+                SNAPSHOT_SCHEMA_VERSION_DEFINITION,
+                null);
+    }
+
+    private boolean shouldCreateNewVersion(DataWorkflow workflow, String incomingSnapshotJson) {
+        if (workflow == null) {
+            return true;
+        }
+        if (workflow.getCurrentVersionId() == null) {
+            return true;
+        }
+        WorkflowVersion currentVersion = workflowVersionMapper.selectById(workflow.getCurrentVersionId());
+        if (currentVersion == null || !StringUtils.hasText(currentVersion.getStructureSnapshot())) {
+            return true;
+        }
+        if (!Objects.equals(currentVersion.getSnapshotSchemaVersion(), SNAPSHOT_SCHEMA_VERSION_DEFINITION)) {
+            return true;
+        }
+        String currentHash = snapshotContentHash(currentVersion.getStructureSnapshot());
+        String incomingHash = snapshotContentHash(incomingSnapshotJson);
+        if (!StringUtils.hasText(currentHash) || !StringUtils.hasText(incomingHash)) {
+            return true;
+        }
+        return !Objects.equals(currentHash, incomingHash);
+    }
+
+    private String snapshotContentHash(String snapshotJson) {
+        if (!StringUtils.hasText(snapshotJson)) {
+            return null;
+        }
+        try {
+            JsonNode node = objectMapper.readTree(snapshotJson);
+            if (node != null && node.isObject()) {
+                ((ObjectNode) node).remove("meta");
+            }
+            String normalized = node != null ? JsonCanonicalizer.canonicalize(node) : snapshotJson.trim();
+            return JsonCanonicalizer.sha256(normalized);
+        } catch (Exception e) {
+            // 非法 JSON 时退化为对原始文本取哈希
+            log.trace("规范化快照 JSON 失败，回退原始文本哈希", e);
+            return JsonCanonicalizer.sha256(snapshotJson.trim());
+        }
+    }
+
+    private List<WorkflowTaskBinding> normalizeTaskBindings(List<WorkflowTaskBinding> tasks) {
+        if (CollectionUtils.isEmpty(tasks)) {
+            return Collections.emptyList();
+        }
+        return tasks;
+    }
+
+    private Long resolveProjectCode(Long requestProjectCode) {
+        if (requestProjectCode != null && requestProjectCode > 0) {
+            return requestProjectCode;
+        }
+        return null;
+    }
+
+    private Long resolveDolphinConfigId(Long requestDolphinConfigId) {
+        if (requestDolphinConfigId != null && requestDolphinConfigId > 0) {
+            return dolphinConfigService.getEnabledConfig(requestDolphinConfigId).getId();
+        }
+        DolphinConfig config = dolphinConfigService.getDefaultConfig();
+        return config != null ? config.getId() : null;
     }
 }
